[chore] 마이페이지, 홈 모듈 메인으로 합치기 (#6)
