--- feature/home/src/main/java/com/readme/android/home/ui/HomeFragment.kt
@@ -1,13 +1,0 @@
-package com.readme.android.home.ui
-
-import android.os.Bundle
-import android.view.View
-import com.readme.android.core.base.BindingFragment
-import com.readme.android.home.R
-import com.readme.android.home.databinding.FragmentHomeBinding
-
-class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home) {
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
