+++ feature/main/build.gradle.kts
@@ -1,0 +1,47 @@
+plugins {
+    id("com.android.library")
+    id("kotlin-parcelize")
+    kotlin("android")
+    kotlin("kapt")
+    id("dagger.hilt.android.plugin")
+}
+
+android {
+    buildFeatures {
+        dataBinding = true
+    }
+}
+
+dependencies {
+    implementation(project(":core"))
+    implementation(project(":domain"))
+    implementation(project(":shared"))
+
+    // Android Core
+    implementation(AndroidXDependencies.coreKtx)
+    implementation(AndroidXDependencies.appCompat)
+    implementation(AndroidXDependencies.constraintLayout)
+    implementation(AndroidXDependencies.coroutines)
+
+    // Material Design
+    implementation(MaterialDesignDependencies.materialDesign)
+
+    // Dagger-Hilt
+    implementation(AndroidXDependencies.hilt)
+    implementation("androidx.appcompat:appcompat:1.4.1")
+    implementation("com.google.android.material:material:1.5.0")
+    implementation("androidx.constraintlayout:constraintlayout:2.1.3")
+    kapt(KaptDependencies.hiltCompiler)
+
+    // Jetpack Lifecycle
+    implementation(AndroidXDependencies.coroutines)
+    implementation(AndroidXDependencies.lifeCycleKtx)
+    implementation(AndroidXDependencies.lifecycleJava8)
+
+    // Jetpack Fragment
+    implementation(AndroidXDependencies.fragment)
+
+    // test
+    implementation(AndroidXDependencies.junit)
+    androidTestImplementation(TestDependencies.androidTest)
+}
