--- feature/mypage/build.gradle.kts
@@ -1,37 +1,0 @@
-plugins {
-    id("com.android.library")
-    kotlin("android")
-    kotlin("kapt")
-    kotlin("plugin.serialization") version Versions.kotlinVersion
-    id("org.jlleitschuh.gradle.ktlint") version Versions.ktlintVersion
-    id("dagger.hilt.android.plugin")
-}
-
-android {
-    buildFeatures {
-        dataBinding = true
-    }
-}
-
-dependencies {
-    implementation(project(":core"))
-    implementation(project(":domain"))
-
-    // Android Core
-    implementation(AndroidXDependencies.coreKtx)
-    implementation(AndroidXDependencies.appCompat)
-    implementation(AndroidXDependencies.constraintLayout)
-    implementation(AndroidXDependencies.legacy)
-    implementation(AndroidXDependencies.coroutines)
-
-    // Material Design
-    implementation(MaterialDesignDependencies.materialDesign)
-
-    // Dagger-Hilt
-    implementation(AndroidXDependencies.hilt)
-    kapt(KaptDependencies.hiltCompiler)
-
-    // test
-    implementation(AndroidXDependencies.junit)
-    androidTestImplementation(TestDependencies.androidTest)
-}
