--- feature/mypage/src/main/java/com/readme/android/mypage/ui/MyPageFragment.kt
@@ -1,7 +1,0 @@
-package com.readme.android.mypage.ui
-
-import com.readme.android.core.base.BindingFragment
-import com.readme.android.mypage.R
-import com.readme.android.mypage.databinding.FragmentMyPageBinding
-
-class MyPageFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_my_page)
