--- feature/mypage/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-
-</resources>
